# docs/swagger.yaml
basePath: /
definitions:
...
  AvailableResource:
    properties:
      id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  ConfigurableResourceSummary:
    properties:
      description:
        type: string
      display_name:
        type: string
      key:
        type: string
    type: object
...
  DiscoveryResult:
    properties:
      resources:
        items:
          $ref: '#/definitions/AvailableResource'
        type: array
    type: object
  JSON:
    additionalProperties: {}
    type: object
...
  agentConnectionResourceSelection:
    properties:
      full_name:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  agentListItem:
    properties:
      attached_skills:
        items:
          $ref: '#/definitions/agentSkillSummary'
        type: array
      avatar_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      instructions:
        type: string
      is_default:
        type: boolean
      last_memory_refreshed_at:
        type: string
      mcp_servers:
        items:
          type: integer
        type: array
      memory_refresh_error:
        type: string
      memory_refresh_status:
        type: string
      model:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/JSON'
      placeholder:
        type: string
      resources:
        $ref: '#/definitions/JSON'
      sandbox:
        $ref: '#/definitions/agentSandboxSummary'
      sandbox_strategy:
        type: string
      sandbox_template_id:
        type: string
      sandbox_tools:
        items:
          type: string
        type: array
      skills:
        $ref: '#/definitions/JSON'
      status:
        type: string
      tools:
        $ref: '#/definitions/JSON'
      triggers:
        items:
          $ref: '#/definitions/agentTriggerResponse'
        type: array
      updated_at:
        type: string
      upgrade_available:
        type: boolean
    type: object
  agentMutationRequest:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      icon:
        type: string
      instructions:
        type: string
      mcp_servers:
        items:
          type: integer
        type: array
      model:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/JSON'
      placeholder:
        type: string
      resources:
        $ref: '#/definitions/JSON'
      sandbox_strategy:
        type: string
      sandbox_template_id:
        type: string
      sandbox_tools:
        items:
          type: string
        type: array
      skills:
        $ref: '#/definitions/JSON'
      tools:
        $ref: '#/definitions/JSON'
    type: object
  agentMutationResponse:
    properties:
      agent:
        $ref: '#/definitions/agentListItem'
    type: object
...
  agentSandboxSummary:
    properties:
      created_at:
        type: string
      error_message:
        type: string
      external_id:
        type: string
      id:
        type: string
      last_active_at:
        type: string
      status:
        type: string
    type: object
...
  agentSkillSummary:
    properties:
      description:
        type: string
      id:
        type: string
      locked:
        type: boolean
      name:
        type: string
      required:
        type: boolean
      source_type:
        type: string
    type: object
  agentTriggerResponse:
    properties:
      conditions: {}
      connection_id:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      instructions:
        type: string
      provider:
        type: string
      secret_set:
        description: |-
          SecretSet indicates whether an HTTP trigger has a shared secret configured.
          True when the trigger requires auth on incoming requests. The secret value
          is never returned.
        type: boolean
      trigger_keys:
        items:
          type: string
        type: array
      trigger_type:
        type: string
    type: object
...
  connectionResponse:
    properties:
      actions_count:
        type: integer
      configurable_resources:
        items:
          $ref: '#/definitions/ConfigurableResourceSummary'
        type: array
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      integration_id:
        type: string
      meta:
        $ref: '#/definitions/JSON'
      nango_connection_id:
        type: string
      org_id:
        type: string
      provider:
        type: string
      provider_config:
        $ref: '#/definitions/JSON'
      revoked_at:
        type: string
      updated_at:
        type: string
      webhook_configured:
        type: boolean
    type: object
...
  createConnectionRequest:
    properties:
      meta:
        $ref: '#/definitions/JSON'
      nango_connection_id:
        type: string
    type: object
...
  createRAGSourceRequest:
    properties:
      access_type:
        type: string
      config:
        $ref: '#/definitions/JSON'
      connection_id:
        type: string
      kind:
        type: string
      name:
        type: string
      perm_sync_freq_seconds:
        type: integer
      prune_freq_seconds:
        type: integer
      refresh_freq_seconds:
        type: integer
    type: object
...
  errorResponse:
    properties:
      error:
        type: string
    type: object
...
  paginatedResponse-agentListItem:
    properties:
      data:
        items:
          $ref: '#/definitions/agentListItem'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
  paginatedResponse-connectionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/connectionResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
...
  ragIndexAttemptResponse:
    properties:
      docs_estimated:
        type: integer
      docs_removed_from_index:
        type: integer
      error_msg:
        type: string
      from_beginning:
        type: boolean
      id:
        type: string
      new_docs_indexed:
        type: integer
      poll_range_end:
        type: string
      poll_range_start:
        type: string
      status:
        type: string
      time_created:
        type: string
      time_started:
        type: string
      time_updated:
        type: string
      total_docs_indexed:
        type: integer
    type: object
...
  ragLatestAttemptStatus:
    properties:
      docs_estimated:
        type: integer
      error_msg:
        type: string
      id:
        type: string
      new_docs_indexed:
        type: integer
      status:
        type: string
      time_started:
        type: string
      time_updated:
        type: string
      total_docs_indexed:
        type: integer
    type: object
  ragListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/ragSourceResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  ragSourceDetailResponse:
    properties:
      access_type:
        type: string
      config:
        items:
          type: integer
        type: array
      connection_id:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      in_repeated_error_state:
        type: boolean
      indexing_start:
        type: string
      kind:
        type: string
      last_pruned:
        type: string
      last_successful_index_time:
        type: string
      last_time_perm_sync:
        type: string
      latest_attempt:
        $ref: '#/definitions/ragLatestAttemptStatus'
      name:
        type: string
      org_id:
        type: string
      perm_sync_freq_seconds:
        type: integer
      prune_freq_seconds:
        type: integer
      recent_attempts:
        items:
          $ref: '#/definitions/ragIndexAttemptResponse'
        type: array
      refresh_freq_seconds:
        type: integer
      status:
        type: string
      total_docs_indexed:
        type: integer
      updated_at:
        type: string
    type: object
  ragSourceResponse:
    properties:
      access_type:
        type: string
      config:
        items:
          type: integer
        type: array
      connection_id:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      in_repeated_error_state:
        type: boolean
      indexing_start:
        type: string
      kind:
        type: string
      last_pruned:
        type: string
      last_successful_index_time:
        type: string
      last_time_perm_sync:
        type: string
      latest_attempt:
        $ref: '#/definitions/ragLatestAttemptStatus'
      name:
        type: string
      org_id:
        type: string
      perm_sync_freq_seconds:
        type: integer
      prune_freq_seconds:
        type: integer
      refresh_freq_seconds:
        type: integer
      status:
        type: string
      total_docs_indexed:
        type: integer
      updated_at:
        type: string
    type: object
...
  updateAgentConnectionResourcesRequest:
    properties:
      resources:
        additionalProperties:
          items:
            $ref: '#/definitions/agentConnectionResourceSelection'
          type: array
        type: object
    type: object
  updateAgentConnectionResourcesResponse:
    properties:
      agent_id:
        type: string
      clone_queued:
        type: boolean
      connection_id:
        type: string
      resources:
        $ref: '#/definitions/JSON'
    type: object
...
  updateRAGSourceRequest:
    properties:
      config:
        $ref: '#/definitions/JSON'
      enabled:
        type: boolean
      indexing_start:
        type: string
      name:
        type: string
      perm_sync_freq_seconds:
        type: integer
      prune_freq_seconds:
        type: integer
      refresh_freq_seconds:
        type: integer
      status:
        type: string
    type: object
...
host: api.dev.usehivy.com
info:
  contact: {}
  description: Proxy runtime for LLM API credentials.
  title: Hivy API
  version: "1.0"
paths:
...
  /v1/agents:
    get:
      description: |-
        Returns all agents in the org with skills (metadata only — no bundle content),
        triggers, and the latest sandbox row.
      parameters:
      - description: Filter by status (draft, active, archived)
        in: query
        name: status
        type: string
      - description: Page size (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paginatedResponse-agentListItem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: List AI agents
      tags:
      - agents
    post:
      consumes:
      - application/json
      description: Creates a user-managed agent definition. Sandbox creation stays
        lazy until the first session.
      parameters:
      - description: Agent create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/agentMutationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/agentMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Create an agent
      tags:
      - agents
  /v1/agents/{id}:
    delete:
      description: Archives an agent when it is not the default Hivy agent and has
        no active sessions.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Archive an agent
      tags:
      - agents
    get:
      description: |-
        Returns one agent in the org with skills (metadata only — no bundle content),
        triggers, and the latest sandbox row.
      parameters:
      - description: Agent agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Get an AI agent
      tags:
      - agents
    patch:
      consumes:
      - application/json
      description: Updates a user-managed agent definition. The default Hivy agent
        cannot be renamed or moved away from always_on.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/agentMutationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Update an agent
      tags:
      - agents
  /v1/agents/{id}/connections/{connectionID}/resources:
    put:
      consumes:
      - application/json
      description: Saves provider resources such as selected GitHub repositories on
        the agent, then queues provider-specific reconciliation when needed.
      parameters:
      - description: Agent UUID
        in: path
        name: id
        required: true
        type: string
      - description: Connection UUID
        in: path
        name: connectionID
        required: true
        type: string
      - description: Selected resources grouped by resource type
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateAgentConnectionResourcesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/updateAgentConnectionResourcesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Save selected connection resources for an agent
      tags:
      - agents
...
  /v1/connections:
    get:
      description: Returns the authenticated user's non-revoked platform integration
        connections.
      parameters:
      - description: Filter by provider
        in: query
        name: provider
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paginatedResponse-connectionResponse'
      security:
      - BearerAuth: []
      summary: List user's connections
      tags:
      - connections
  /v1/connections/{id}:
    delete:
      description: Revokes a user's platform integration connection and removes it
        from Nango.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Disconnect an connection
      tags:
      - connections
...
  /v1/connections/{id}/resources/{type}:
    get:
      description: Fetches available resources of a specific type from the provider
        API. For example, list all repositories for a GitHub connection.
      parameters:
      - description: In-Connection ID
        in: path
        name: id
        required: true
        type: string
      - description: Resource type (e.g., repository, project)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DiscoveryResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: List available resources for a connection
      tags:
      - connections
...
  /v1/integrations/{id}/connections:
    post:
      consumes:
      - application/json
      description: Stores a connection after the OAuth flow completes via Nango.
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      - description: Connection details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/connectionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Create an connection
      tags:
      - connections
...
  /v1/rag/sources:
    get:
      description: Returns the org's RAG sources. Supports pagination and optional
        status / kind filters.
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by kind
        in: query
        name: kind
        type: string
      - description: Page number (0-indexed)
        in: query
        name: page
        type: integer
      - description: Page size, max 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ragListResponse'
      security:
      - BearerAuth: []
      summary: List RAG sources
      tags:
      - rag
    post:
      consumes:
      - application/json
      description: Creates a new RAG source that the scheduler will pick up on the
        next tick. Kind=integration requires a valid connection_id pointing at an
        integration whose supports_rag_source flag is true. Refresh / prune / perm-sync
        frequencies are validated against per-org minimums.
      parameters:
      - description: Source definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createRAGSourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ragSourceResponse'
      security:
      - BearerAuth: []
      summary: Create a RAG source
      tags:
      - rag
  /v1/rag/sources/{id}:
    delete:
      description: Hard-deletes the source row. Postgres-side rows (attempts, documents,
        sync state, ACLs) cascade. Vector store entries are reaped later by the prune
        loop.
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a RAG source
      tags:
      - rag
    get:
      description: Returns one RAG source by ID with the last 5 index attempts inlined.
        404 on cross-org access by design.
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ragSourceDetailResponse'
      security:
      - BearerAuth: []
      summary: Get a RAG source
      tags:
      - rag
    patch:
      consumes:
      - application/json
      description: Patches mutable fields (name, status pause/resume, refresh / prune
        / perm-sync frequencies, indexing_start floor). Only ACTIVE ↔ PAUSED transitions
        are client-settable; DELETING is server-managed.
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      - description: Patch payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateRAGSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ragSourceResponse'
      security:
      - BearerAuth: []
      summary: Update a RAG source
      tags:
      - rag
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token (JWT or API key).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"